testing/web-platform/tests/css/css-backgrounds/parsing/background-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Backgrounds and Borders Module Level 3: parsing background with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-backgrounds/#background">
<meta name="assert" content="background supports the full grammar '<bg-layer># , <final-bg-layer>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/parsing-testcommon.js"></script>
</head>
<body>
<script>
// For <final-bg-layer>, implementations serialize <‘background-color’> after <bg-image>, contrary to spec.
// Safari removes quotes.
test_valid_value("background", '1px 2px / 3px 4px space round fixed border-box padding-box, red url("https://example.com/")', ['1px 2px / 3px 4px space round fixed border-box padding-box, url("https://example.com/") red', '1px 2px / 3px 4px space round fixed border-box padding-box, url(https://example.com/) red']);
</script>
</body>
</html>
